split dependencies out of build script
