+++ modules/secure-upload/src/build-lib/README.md
@@ -1,0 +1,5 @@
+# build-lib
+
+build-lib is a folder where we store all of the libraries used in the build
+script. We wanted to separate that code out from the code and libraries that
+get in the module itself.
